Change types of identity properties to unsigned

+++ Marathon.Domain/Entities/Charity.cs
@@ -6,14 +6,14 @@ namespace Marathon.Domain.Entities
     /// <summary>
     /// Details of the <see cref="Charity"/>
     /// </summary>
-    public sealed class Charity : IEntity<long>
+    public sealed class Charity : IEntity<uint>
     {
         public Charity()
         {
             SignUps = new HashSet<MarathonSignUp>();
         }
 
-        public long Id { get; set; }
+        public uint Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public byte[] Logo { get; set; }
